+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased - YYYY-MM-DD
 
 ### Changed
+
 - Update dependencies;
 
 ## 2.0.1-rc.1 - 2022-09-28