+++ CHANGELOG.md
@@ -30,6 +30,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update dependencies;
 - Return `NetworkInfoDto` instead of `NetworkInfo` in message_interface;
 
+## 2.0.1-rc.2 - 2022-09-29
+
+### Added
+
+- `NetworkInfoDto`;
+
+### Changed
+
+- Update dependencies;
+- Return `NetworkInfoDto` instead of `NetworkInfo` in message_interface;
+- Use `#[cfg(target_family = "wasm")]` instead of `#[cfg(feature = "wasm")]`;
+
 ## 2.0.1-rc.1 - 2022-09-28
 
 Re-release as RC.